requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
binaryornot==0.4.4
bitarray==2.7.3
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
cryptography==40.0.1
cytoolz==0.12.1
Django==4.1.7
django-browser-reload==1.7.0
dune-client==1.0.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
filelock==3.10.7
frozenlist==1.3.3
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
lru-dict==1.1.8
Markdown==3.3.4
MarkupSafe==2.1.2
martor==1.6.26
multidict==6.0.4
ndjson==0.3.1
numpy==1.24.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
parsimonious==0.9.0
protobuf==4.22.3
pyarrow==10.0.1
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
six==1.16.0
snowflake-connector-python==3.0.2
sqlparse==0.4.3
text-unidecode==1.3
toolz==0.12.0
types-python-dateutil==2.8.19.12
types-PyYAML==6.0.12.9
types-requests==2.28.11.17
types-urllib3==1.26.25.10
typing_extensions==4.5.0
urllib3==1.26.15
web3==6.2.0
webencodings==0.5.1
websockets==11.0.2
yarl==1.9.1
